[REFACTOR] #214 - RePackaging
